requirements.txt
# Core dependencies
pandas>=1.3.0
numpy>=1.20.0
requests>=2.25.0
beautifulsoup4>=4.9.0
python-dotenv>=0.19.0

# Data visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.3.0

# NLP and AI
openai>=1.0.0
scikit-learn>=1.0.0

# Web scraping
lxml>=4.6.0

# Dashboard
flask>=2.0.0